+++ picodi/__init__.py
@@ -2,7 +2,6 @@ from picodi._picodi import (
     Provide,
     init_resources,
     inject,
-    make_dependency,
     registry,
     resource,
     shutdown_resources,
@@ -11,7 +10,6 @@ from picodi._picodi import (
 __all__ = [
     "init_resources",
     "inject",
-    "make_dependency",
     "Provide",
     "registry",
     "resource",